feat : 대기페이지 레이아웃 재배치 및 고양이 선택 영역 분리
