+++ client/src/pages/WaitingRoomPage.tsx
@@ -7,25 +7,29 @@ import modalState from '@store/modal';
 import userState from '@store/user';
 import roomState from '@store/room';
 
+import RoundSquareButton from '@atoms/RoundSquareButton';
+import RuleBookButton from '@atoms/RuleBookButton';
 import WaitingRoomChat from '@molecules/WaitingRoomChat';
 import WaitingRoomUsers from '@molecules/WaitingRoomUsers';
-import RoundSquareButton from '@atoms/RoundSquareButton';
-import RuleBookButton from '@atoms/RuleBookButton'
+import CatSelector from '@organisms/CatSelector';
 
-interface ButtonProps {
-  isOpen: boolean;
-}
 const Container = styled.div`
   width: 100%;
   height: 100%;
   display: flex;
   flex-direction: column;
 `;
+
 const Header = styled.div`
   text-align: center;
   padding-left: 20px;
   border-bottom: 1px solid gray;
 `;
+
+const UserContainer = styled.div`
+  width: 240px;
+`;
+
 const Content = styled.section`
   text-align: center;
   display: flex;
@@ -35,37 +39,28 @@ const Content = styled.section`
 const Chat = styled.div`
   padding: 10px;
   margin-left: 20px;
-  flex-basis: 300px;
-  flex-grow: 1;
+  width: 360px;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   border-left: 1px solid gray;
   border-right: 1px solid gray;
 `;
-const Character = styled.div`
+
+const WaitingZone = styled.div`
   display: flex;
-  flex-grow: 9;
+  flex: 1;
   flex-direction: column;
   align-items: center;
   padding: 10px;
 `;
+
 const Footer = styled.div`
   padding: 20px;
   border-top: 1px solid gray;
   display: flex;
   justify-content: flex-end;
 `;
-
-const SelectCharacter = styled.div`
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-`;
-const CharacterImg = styled.img`
-  width: 400px;
-`;
-const SelectCharacterButton = styled.button<ButtonProps>``;
 
 const RuleBookContainer = styled.div`
   position: fixed;
@@ -75,7 +70,6 @@ const RuleBookContainer = styled.div`
 
 const WaitingRoomPage = () => {
   const history = useHistory();
-  const [isOpen, setIsOpen] = useState(false);
   const [modal, setModal] = useRecoilState(modalState);
   const [user, setUser] = useRecoilState(userState);
   const [room, setRoom] = useRecoilState(roomState);
@@ -84,22 +78,10 @@ const WaitingRoomPage = () => {
     console.log(user, room);
   }, []);
 
-  const selectCharacter = () => {
-    setIsOpen(true);
-    setModal('SelectCharacterModal');
-  };
-
-  useEffect(() => {
-    if (modal !== 'SelectCharacterModal') {
-      setIsOpen(false);
-    }
-  }, [modal]);
-
   function goRobby() {
-    let selected = confirm('대기방을 나갑니다.');
+    let selected = confirm('대기방을 나가시겠습니까?');
 
     if (selected) {
-      alert('로비로 이동합니다.');
       history.push('/');
     }
   }
@@ -118,26 +100,16 @@ const WaitingRoomPage = () => {
         <h1>Waiting Room Page</h1>
       </Header>
       <Content>
-        <WaitingRoomUsers></WaitingRoomUsers>
+        <UserContainer>
+          <h2>캐릭터 선택</h2>
+          <CatSelector />
+          <WaitingRoomUsers />
+        </UserContainer>
         <Chat>
-          <div>
-            <h2>채팅</h2>
-          </div>
-          <div>
-            <WaitingRoomChat></WaitingRoomChat>
-          </div>
+          <h2>채팅</h2>
+          <WaitingRoomChat />
         </Chat>
-        <Character>
-          <div>
-            <h2>캐릭터 선택</h2>
-          </div>
-          <SelectCharacter>
-            <CharacterImg alt="캐릭터" />
-            <SelectCharacterButton isOpen={isOpen} onClick={selectCharacter}>
-              캐릭터 선택
-            </SelectCharacterButton>
-          </SelectCharacter>
-        </Character>
+        <WaitingZone></WaitingZone>
       </Content>
       <Footer>
         <RoundSquareButton variant="yellow" size="lg" onClick={setReady}>
@@ -146,7 +118,6 @@ const WaitingRoomPage = () => {
         <RoundSquareButton variant="yellow" size="lg" onClick={getHelp}>
           도움말
         </RoundSquareButton>
-
         <RoundSquareButton variant="yellow" size="lg" onClick={goRobby}>
           로비로 돌아가기
         </RoundSquareButton>
